Refactor shared non-coin specific functions into utils module
Share non-coin specific functions like isValidHex() and isValidAddress() are moved to utils/validate.ts, and associated hashtypes model and interface HashType are moved to utils/hash.ts for decoupling purpose. Any coins added in future can share shared functions and base58 policy can be extended more easily

The shared functions are called as Validator module in each file when imported
