+++ src/coin/xtz/transactionBuilder.ts
@@ -4,6 +4,7 @@ import { BaseCoin as CoinConfig } from '@bitgo/statics';
 import { SigningError, BuildTransactionError } from '../baseCoin/errors';
 import { BaseKey } from '../baseCoin/iface';
 import { BaseTransactionBuilder, TransactionType } from '../baseCoin';
+import * as Validator from '../../utils/validate';
 import {
   genericMultisigOriginationOperation,
   multisigTransactionOperation,
@@ -14,14 +15,8 @@ import { Address } from './address';
 import { Transaction } from './transaction';
 import { KeyPair } from './keyPair';
 import { Fee, IndexedData, IndexedSignature, Key, Operation, OriginationOp, RevealOp, TransactionOp } from './iface';
-import {
-  isValidAddress,
-  isValidBlockHash,
-  DEFAULT_GAS_LIMIT,
-  DEFAULT_STORAGE_LIMIT,
-  isValidOriginatedAddress,
-  isValidPublicKey,
-} from './utils';
+import { DEFAULT_GAS_LIMIT, DEFAULT_STORAGE_LIMIT } from './utils';
+
 import * as Utils from './utils';
 import { TransferBuilder } from './transferBuilder';
 
@@ -194,7 +189,7 @@ export class TransactionBuilder extends BaseTransactionBuilder {
    * @param {string} blockId A block hash to use as branch reference
    */
   branch(blockId: string): void {
-    if (!isValidBlockHash(blockId)) {
+    if (!Validator.isValidBlockHash(blockId)) {
       throw new BuildTransactionError('Invalid block hash ' + blockId);
     }
     this._blockHeader = blockId;
@@ -277,7 +272,7 @@ export class TransactionBuilder extends BaseTransactionBuilder {
     if (this._publicKeyToReveal) {
       throw new BuildTransactionError('Public key to reveal already set: ' + this._publicKeyToReveal);
     }
-    if (!isValidPublicKey(publicKey)) {
+    if (!Validator.isValidPublicKey(publicKey)) {
       throw new BuildTransactionError('Invalid public key: ' + publicKey);
     }
     if (new KeyPair({ pub: publicKey }).getAddress() !== this._sourceAddress) {
@@ -311,7 +306,7 @@ export class TransactionBuilder extends BaseTransactionBuilder {
     if (this._walletOwnerPublicKeys.length >= DEFAULT_M) {
       throw new BuildTransactionError('A maximum of ' + DEFAULT_M + ' owners can be set for a multisig wallet');
     }
-    if (!isValidPublicKey(publicKey)) {
+    if (!Validator.isValidPublicKey(publicKey)) {
       throw new BuildTransactionError('Invalid public key: ' + publicKey);
     }
     if (this._walletOwnerPublicKeys.includes(publicKey)) {
@@ -405,7 +400,7 @@ export class TransactionBuilder extends BaseTransactionBuilder {
     for (let i = 0; i < this._transfers.length; i++) {
       const transfer = this._transfers[i].build();
       let transactionOp;
-      if (isValidOriginatedAddress(transfer.from)) {
+      if (Validator.isValidOriginatedAddress(transfer.from)) {
         // Offline transactions may not have the data to sign
         const signatures = transfer.dataToSign ? await this.getSignatures(transfer.dataToSign) : [];
         transactionOp = multisigTransactionOperation(
@@ -449,7 +444,7 @@ export class TransactionBuilder extends BaseTransactionBuilder {
 
   /** @inheritdoc */
   validateAddress(address: Address): void {
-    if (!isValidAddress(address.address)) {
+    if (!Validator.isValidAddress(address.address)) {
       throw new BuildTransactionError('Invalid address ' + address.address);
     }
   }
